chore: make more explicit that the WaitKubernetesAPIServer is tied to the management package
