+++ internal/cmd/manager/instance/pgbasebackup/cmd.go
@@ -31,7 +31,6 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/external"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/log"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/postgres"
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/resources"
 )
 
 // CloneInfo is the structure containing all the information needed
@@ -51,7 +50,7 @@ func NewCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use: "pgbasebackup",
 		PreRunE: func(cmd *cobra.Command, args []string) error {
-			return resources.WaitKubernetesAPIServer(cmd.Context(), ctrl.ObjectKey{
+			return management.WaitKubernetesAPIServer(cmd.Context(), ctrl.ObjectKey{
 				Name:      clusterName,
 				Namespace: namespace,
 			})
